+++ refacFabela/src/main/java/com/refacFabela/repository/AbonoVentaIdRepository.java
@@ -1,5 +1,6 @@
 package com.refacFabela.repository;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -19,9 +20,9 @@ public interface AbonoVentaIdRepository extends JpaRepository<TwAbono, Long> {
 	@Query("Select e from TwAbono e where e.twCaja.nId= :nIdCaja ")
 	public List<TwAbono> obtenerAbonosCaja(Long nIdCaja);
 	 @Query(value = "Select truncate(ifnull(sum(e.n_abono),0),2) from tw_abonos e where e.n_idCaja=:nIdCaja and e.n_idFormaPago=1 ",   nativeQuery = true) 
-	public Double TotalAbonoEfectivo(Long nIdCaja);
+	public BigDecimal TotalAbonoEfectivo(Long nIdCaja);
 	 @Query(value = "select truncate(ifnull(sum(e.n_abono),0),2) from tw_abonos e where e.n_idCaja=:nIdCaja and (e.n_idFormaPago=2 or e.n_idFormaPago=3  or e.n_idFormaPago=4  or e.n_idFormaPago=18)",   nativeQuery = true) 
-	public Double TotalAbonoElectronico(Long nIdCaja);
+	public BigDecimal TotalAbonoElectronico(Long nIdCaja);
 	
 
 }
