+++ refacFabela/src/main/java/com/refacFabela/repository/TrVentaCobroRepository.java
@@ -1,5 +1,6 @@
 package com.refacFabela.repository;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -20,10 +21,10 @@ public interface TrVentaCobroRepository extends JpaRepository<TrVentaCobro, Long
 
     
     @Query(value = "select truncate(ifnull(sum(c.n_monto),0),2)  from tr_venta_cobro c where c.n_id_caja=:nIdCaja and n_id_forma_pago=1",   nativeQuery = true) 
-    public Double TotalPagoEfectivoVenta(Long nIdCaja);
+    public BigDecimal TotalPagoEfectivoVenta(Long nIdCaja);
     
     @Query(value = "select truncate(ifnull(sum(c.n_monto),0),2) from tr_venta_cobro c where c.n_id_caja=:nIdCaja and (n_id_forma_pago=2 or  n_id_forma_pago=3 or  n_id_forma_pago=4 or  n_id_forma_pago=18)",   nativeQuery = true) 
-    public Double TotalPagoElectronico(Long nIdCaja);
+    public BigDecimal TotalPagoElectronico(Long nIdCaja);
     
     public void deleteBynIdVenta(Long nIdVenta);
